Address review: single-probe half-open breaker, explicit outage signal, audit export hardening
- EmbeddingClient: make HALF_OPEN a deterministic single-probe state (reject
  concurrent callers by state, not a time deadline), keep isAvailable() false
  while a probe is in flight, and record permit-acquisition failures so a
  promoted probe always resolves. Adds a latch-based single-probe test.
- VectorDocBuilder.fromEntity: throw EmbeddingUnavailableException on outage
  instead of returning an empty list callers cannot distinguish from
  "no chunks" (which would delete existing vectors).
- OpenSearchVectorService: drop the racy chunksEmptyFromOutage guard; catch the
  outage exception so replaceChunks never deletes chunks during a provider outage.
- OpenSearchBulkSink: skip the embedding call during an outage so a transient
  provider outage indexes without embeddings (self-heals next reindex) instead
  of failing every entity in the reindex.
- SearchClusterMetrics.boundConcurrentRequestsToMemory: compute the memory cap
  in long and cast after Math.min so a tiny payload can't overflow to negative
  concurrency. Adds a regression test.
- AuditLogResource: validate the export jobId is a server-generated UUID before
  touching the spool path (path-traversal guard) and delete a partial spool on
  export failure.

+++ openmetadata-service/src/main/java/org/openmetadata/service/search/SearchClusterMetrics.java
@@ -440,8 +440,8 @@ public class SearchClusterMetrics {
   static int boundConcurrentRequestsToMemory(
       int concurrentRequests, long maxPayloadSizeBytes, long offHeapBudgetBytes) {
     long safeBytes = (long) (offHeapBudgetBytes * INFLIGHT_MEMORY_SAFETY_FACTOR);
-    int maxByMemory = (int) Math.max(1, safeBytes / Math.max(1, maxPayloadSizeBytes));
-    int bounded = Math.min(concurrentRequests, maxByMemory);
+    long maxByMemory = Math.max(1, safeBytes / Math.max(1, maxPayloadSizeBytes));
+    int bounded = (int) Math.min(concurrentRequests, maxByMemory);
     if (bounded < concurrentRequests) {
       LOG.info(
           "Capping concurrent bulk requests {} -> {} to fit off-heap budget {} MB (payload {} MB); "
